refactor(commands): extract template tests to sibling file
